+++ pyincore/utils/analysisutil.py
@@ -223,6 +223,48 @@ class AnalysisUtil:
         return result_dataset_id
 
     @staticmethod
+    def adjust_damage_for_liquefaction(limit_state_probabilities, ground_failure_probabilities):
+        """Adjusts building damage probability based on liquefaction ground failure probability
+        with the liq_dmg, we know that it is 3 values, the first two are the same.
+        The 3rd might be different.
+        We always want to apply the first two to all damage states except the highest.
+
+        Args:
+            limit_state_probabilities (obj): Limit state probabilities.
+            ground_failure_probabilities (list): Ground failure probabilities.
+
+        Returns:
+            OrderedDict: Adjusted limit state probability.
+
+        """
+        keys = list(limit_state_probabilities.keys())
+        adjusted_limit_state_probabilities = collections.OrderedDict()
+
+        for i in range(len(keys)):
+            # check and see...if we are trying to use the last ground failure
+            # number for something other than the
+            # last limit-state-probability, then we should use the
+            # second-to-last probability of ground failure instead.
+
+            if i > len(ground_failure_probabilities) - 1:
+                prob_ground_failure = ground_failure_probabilities[len(ground_failure_probabilities) - 2]
+            else:
+                prob_ground_failure = ground_failure_probabilities[i]
+
+            adjusted_limit_state_probabilities[keys[i]] = \
+                limit_state_probabilities[keys[i]] + prob_ground_failure \
+                - limit_state_probabilities[keys[i]] * prob_ground_failure
+
+        # the final one is the last of limitStates should match with the last of ground failures
+        j = len(limit_state_probabilities) - 1
+        prob_ground_failure = ground_failure_probabilities[-1]
+        adjusted_limit_state_probabilities[keys[j]] = \
+            limit_state_probabilities[keys[j]] \
+            + prob_ground_failure - limit_state_probabilities[keys[j]] * prob_ground_failure
+
+        return adjusted_limit_state_probabilities
+
+    @staticmethod
     def adjust_limit_states_for_pgd(limit_states, pgd_limit_states):
         try:
             adj_limit_states = collections.OrderedDict()
@@ -554,3 +596,31 @@ class AnalysisUtil:
 
         """
         return any("-9999" in str(val) for val in hazard_vals)
+
+    @staticmethod
+    def get_discretized_restoration(restoration_table):
+        """Converts discretized restoration times into a dictionary
+
+        Args:
+            restoration_table(list):
+        Returns:
+            dict: Grouped discretized restoration by inventory class { 'EPPL': {day1: [100, 50, 9, 4, 3], day3: [100,
+            100, 50, 13, 4], etc }
+        """
+
+        restoration_map = {}
+        for restoration_row in restoration_table:
+            classification = restoration_row['Classification']
+            if classification in restoration_map:
+                class_restoration = restoration_map[classification]
+            else:
+                class_restoration = {"day1": [], "day3": [], "day7": [], "day30": [], "day90": []}
+                restoration_map[classification] = class_restoration
+
+            class_restoration["day1"].append(float(restoration_row["day1"]))
+            class_restoration["day3"].append(float(restoration_row["day3"]))
+            class_restoration["day7"].append(float(restoration_row["day7"]))
+            class_restoration["day30"].append(float(restoration_row["day30"]))
+            class_restoration["day90"].append(float(restoration_row["day90"]))
+
+        return restoration_map