+++ tests/test_format.py
@@ -65,6 +65,12 @@ paths = [
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/utils/test_networkutil.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/models/test_networkdataset.py'),
     os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/test_analysis.py')
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/waterfacilityrestoration/waterfacilityrestoration.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/waterfacilityrestoration/waterfacilityrestorationutil.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER,
+                 'tests/pyincore/analyses/waterfacilityrestoration/test_waterfacilityrestoration.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'pyincore/analyses/socialvulnerability/socialvulnerability.py'),
+    os.path.join(PYINCORE_ROOT_FOLDER, 'tests/pyincore/analyses/socialvulnerability/test_socialvulnerability.py')
 ]